requirements.txt
# LangChain et composants
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-openai>=0.0.5

# Modèles de langage
openai>=1.10.0
ollama>=0.1.4
tiktoken>=0.5.2

# Base de données vectorielle
chromadb>=0.4.22
sentence-transformers>=2.3.1

# API et Web
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.26.0

# Interface utilisateur
tkinter  # Inclus dans Python standard
PyQt6>=6.6.1  # Optionnel pour GUI avancée

# Traitement des données
pydantic>=2.6.0
numpy>=1.24.0
pandas>=2.1.0

# Sécurité et logging
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.0
python-multipart>=0.0.6
logging>=0.5.1.2

# Tests
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Utilitaires
python-dateutil>=2.8.2
pytz>=2024.1
tqdm>=4.66.0 